Use a generic name for word limit error messages
Previously the error message was generated using the name of the error
message, which was generated by the word limit itself.  This meant that
new error names had to be generated whenever word limits were changed.
There's no need for this now because the error messages are in the yml
files.